+++ frontend/components/views/QueryView/QueryTabs.tsx
@@ -1,29 +1,10 @@
 import React, { useState } from 'react';
-import { QueryData } from '../../../types';
+import { Container } from '@material-ui/core';
+import { QueryData, ValidTabs } from '../../../types';
 
+import TabSelector from './TabSelector';
 import QueryResults from './QueryResults';
 import QueryPlan from './QueryPlan';
-
-type ValidTabs = 'Results' | 'Execution Plan';
-
-interface TabSelectorProps {
-  selectedTab: ValidTabs;
-  select: (tab: ValidTabs) => void;
-}
-
-const TabSelector = ({ selectedTab, select }: TabSelectorProps) => {
-  const tabs: ValidTabs[] = ['Results', 'Execution Plan'];
-
-  return (
-    <div>
-      {tabs.map((tab: ValidTabs) => (
-        <button type="button" onClick={() => select(tab)} key={`querytab_${tab}`}>
-          {`${tab}${selectedTab === tab ? ' <' : ''}`}
-        </button>
-      ))}
-    </div>
-  );
-};
 
 interface QueryTabsProps {
   results: QueryData['returnedRows'];
@@ -39,12 +20,12 @@ const QueryTabs = ({ results, executionPlan }: QueryTabsProps) => {
         selectedTab={selectedTab}
         select={(tab: ValidTabs) => setSelectedTab(tab)}
       />
-      <div>
+      <Container>
         {selectedTab === 'Results' ? <QueryResults results={results} /> : null}
         {selectedTab === 'Execution Plan' ? (
           <QueryPlan executionPlan={executionPlan} />
         ) : null}
-      </div>
+      </Container>
     </div>
   );
 };
